// src/domains/service-clients/order/proto/generated/order_service.ts
export interface OrderItemsData {
  courseId: string;
  price: number;
}

export interface PaymentDetailsData {
  paymentId: string;
  provider: string;
  providerOrderId?: string | undefined;
  paymentStatus: string;
  updatedAt: string;
}

export interface MoneyData {
  total: number;
  currency: string;
  discount: number;
  salesTax?: number | undefined;
  subTotal: number;
}

export interface OrderData {
  id: string;
  userId: string;
  items: OrderItemsData[];
  paymentDetails?: PaymentDetailsData | undefined;
  amount: MoneyData | undefined;
  status: string;
  createdAt: string;
  updatedAt: string;
}
// ...
export const OrderItemsData: MessageFns<OrderItemsData> = {
  encode(message: OrderItemsData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.courseId !== "") {
      writer.uint32(10).string(message.courseId);
    }
    if (message.price !== 0) {
      writer.uint32(21).float(message.price);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OrderItemsData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOrderItemsData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.courseId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 21) {
            break;
          }

          message.price = reader.float();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): OrderItemsData {
    return {
      courseId: isSet(object.courseId) ? globalThis.String(object.courseId) : "",
      price: isSet(object.price) ? globalThis.Number(object.price) : 0,
    };
  },

  toJSON(message: OrderItemsData): unknown {
    const obj: any = {};
    if (message.courseId !== "") {
      obj.courseId = message.courseId;
    }
    if (message.price !== 0) {
      obj.price = message.price;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<OrderItemsData>, I>>(base?: I): OrderItemsData {
    return OrderItemsData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OrderItemsData>, I>>(object: I): OrderItemsData {
    const message = createBaseOrderItemsData();
    message.courseId = object.courseId ?? "";
    message.price = object.price ?? 0;
    return message;
  },
};

function createBasePaymentDetailsData(): PaymentDetailsData {
  return { paymentId: "", provider: "", providerOrderId: undefined, paymentStatus: "", updatedAt: "" };
}

export const PaymentDetailsData: MessageFns<PaymentDetailsData> = {
  encode(message: PaymentDetailsData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.paymentId !== "") {
      writer.uint32(10).string(message.paymentId);
    }
    if (message.provider !== "") {
      writer.uint32(18).string(message.provider);
    }
    if (message.providerOrderId !== undefined) {
      writer.uint32(26).string(message.providerOrderId);
    }
    if (message.paymentStatus !== "") {
      writer.uint32(34).string(message.paymentStatus);
    }
    if (message.updatedAt !== "") {
      writer.uint32(42).string(message.updatedAt);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PaymentDetailsData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePaymentDetailsData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.paymentId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.provider = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.providerOrderId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.paymentStatus = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.updatedAt = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PaymentDetailsData {
    return {
      paymentId: isSet(object.paymentId) ? globalThis.String(object.paymentId) : "",
      provider: isSet(object.provider) ? globalThis.String(object.provider) : "",
      providerOrderId: isSet(object.providerOrderId) ? globalThis.String(object.providerOrderId) : undefined,
      paymentStatus: isSet(object.paymentStatus) ? globalThis.String(object.paymentStatus) : "",
      updatedAt: isSet(object.updatedAt) ? globalThis.String(object.updatedAt) : "",
    };
  },

  toJSON(message: PaymentDetailsData): unknown {
    const obj: any = {};
    if (message.paymentId !== "") {
      obj.paymentId = message.paymentId;
    }
    if (message.provider !== "") {
      obj.provider = message.provider;
    }
    if (message.providerOrderId !== undefined) {
      obj.providerOrderId = message.providerOrderId;
    }
    if (message.paymentStatus !== "") {
      obj.paymentStatus = message.paymentStatus;
    }
    if (message.updatedAt !== "") {
      obj.updatedAt = message.updatedAt;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PaymentDetailsData>, I>>(base?: I): PaymentDetailsData {
    return PaymentDetailsData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PaymentDetailsData>, I>>(object: I): PaymentDetailsData {
    const message = createBasePaymentDetailsData();
    message.paymentId = object.paymentId ?? "";
    message.provider = object.provider ?? "";
    message.providerOrderId = object.providerOrderId ?? undefined;
    message.paymentStatus = object.paymentStatus ?? "";
    message.updatedAt = object.updatedAt ?? "";
    return message;
  },
};

function createBaseMoneyData(): MoneyData {
  return { total: 0, currency: "", discount: 0, salesTax: undefined, subTotal: 0 };
}

export const MoneyData: MessageFns<MoneyData> = {
  encode(message: MoneyData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.total !== 0) {
      writer.uint32(13).float(message.total);
    }
    if (message.currency !== "") {
      writer.uint32(18).string(message.currency);
    }
    if (message.discount !== 0) {
      writer.uint32(29).float(message.discount);
    }
    if (message.salesTax !== undefined) {
      writer.uint32(53).float(message.salesTax);
    }
    if (message.subTotal !== 0) {
      writer.uint32(37).float(message.subTotal);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MoneyData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMoneyData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 13) {
            break;
          }

          message.total = reader.float();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.currency = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.discount = reader.float();
          continue;
        }
        case 6: {
          if (tag !== 53) {
            break;
          }

          message.salesTax = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.subTotal = reader.float();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MoneyData {
    return {
      total: isSet(object.total) ? globalThis.Number(object.total) : 0,
      currency: isSet(object.currency) ? globalThis.String(object.currency) : "",
      discount: isSet(object.discount) ? globalThis.Number(object.discount) : 0,
      salesTax: isSet(object.salesTax) ? globalThis.Number(object.salesTax) : undefined,
      subTotal: isSet(object.subTotal) ? globalThis.Number(object.subTotal) : 0,
    };
  },

  toJSON(message: MoneyData): unknown {
    const obj: any = {};
    if (message.total !== 0) {
      obj.total = message.total;
    }
    if (message.currency !== "") {
      obj.currency = message.currency;
    }
    if (message.discount !== 0) {
      obj.discount = message.discount;
    }
    if (message.salesTax !== undefined) {
      obj.salesTax = message.salesTax;
    }
    if (message.subTotal !== 0) {
      obj.subTotal = message.subTotal;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MoneyData>, I>>(base?: I): MoneyData {
    return MoneyData.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MoneyData>, I>>(object: I): MoneyData {
    const message = createBaseMoneyData();
    message.total = object.total ?? 0;
    message.currency = object.currency ?? "";
    message.discount = object.discount ?? 0;
    message.salesTax = object.salesTax ?? undefined;
    message.subTotal = object.subTotal ?? 0;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}